Add files via upload
Human longevity is an important but difficult goal due to the extreme complication of human body. If people could repeatedly transfer their consciousness from old bodies to new ones, their lifespan might be prolonged extremely. However, there are several difficulties which prevent the achievement of such a technique. In this paper, we propose an approach that can transfer human consciousness indirectly, and avoid or significantly relieve those difficulties. The strategy of our approach is to integrate the consciousnesses of two bodies (old and new) into a unique consciousness. By doing so, the consciousness is extended to the new body. After that, the consciousness shrinks and remains in the new body when the old body dies. This may also apply to integrating more than two bodies of human, or animal, or even across different species. After investigating thousands of literatures, we find quite strong evidence (related discoveries and technologies) that can support the proposed approach and its advantages. Beside prolonging human lifespan, the approach could also have other meaningful applications.